refactor: move incomingwebhookdata to types
